move most files to main app_check folder
