+++ app_check/src/include/firebase/app_check.h
@@ -18,11 +18,82 @@ namespace app_check {
 #ifndef FIREBASE_APP_CHECK_SRC_INCLUDE_FIREBASE_APP_CHECK_H_
 #define FIREBASE_APP_CHECK_SRC_INCLUDE_FIREBASE_APP_CHECK_H_
 
-#include "firebase/app_check/app_check_provider.h"
-#include "firebase/app_check/app_check_provider_factory.h"
-#include "firebase/app_check/firebase_app_check.h"
+#include "app_check/src/app_check_provider.h"
+#include "app_check/src/app_check_provider_factory.h"
+
+/// Struct to hold tokens emitted by the Firebase App Check service which are
+/// minted upon a successful application verification. These tokens are the
+/// federated output of a verification flow, the structure of which is
+/// independent of the mechanism by which the application was verified.
+struct AppCheckToken {
+  /// A Firebase App Check token.
+  std::string token;
+
+  /// A Firebase App Check token expiration date in the device local time.
+  int64_t expire_time_millis;
+}
+
+class AppCheckListener {
+  virtual ~AppCheckListener();
+  /**
+   * This method gets invoked on the UI thread on changes to the token state.
+   * Does not trigger on token expiry.
+   */
+  virtual void OnAppCheckTokenChanged(const AppCheckToken& token) = 0;
+}
+
+class AppCheck {
+ public:
+  /**
+   * Gets the instance of {@code AppCheck} associated with the given {@link
+   * FirebaseApp} instance.
+   */
+  static AppCheck* GetInstance(::firebase::App* app);
+
+  /**
+   * Installs the given {@link AppCheckProviderFactory}, overwriting any that
+   * were previously associated with this {@code AppCheck} instance. Any {@link
+   * AppCheckTokenListener}s attached to this {@code AppCheck} instance will be
+   * transferred from existing factories to the newly installed one.
+   *
+   * <p>Automatic token refreshing will only occur if the global {@code
+   * isDataCollectionDefaultEnabled} flag is set to true. To allow automatic
+   * token refreshing for Firebase App Check without changing the {@code
+   * isDataCollectionDefaultEnabled} flag for other Firebase SDKs, use {@link
+   * #setAppCheckProviderFactory(AppCheckProviderFactory, bool)} instead or call
+   * {@link #setTokenAutoRefreshEnabled(bool)} after installing the {@code
+   * factory}.
+   *
+   * This method should be called before initializing the Firebase App.
+   */
+  static void SetAppCheckProviderFactory(
+      const AppCheckProviderFactory& factory) = 0;
+
+  /** Sets the {@code isTokenAutoRefreshEnabled} flag. */
+  void SetTokenAutoRefreshEnabled(bool is_token_auto_refresh_enabled) =
+      0;
+
+  /**
+   * Requests a Firebase App Check token. This method should be used ONLY if you
+   * need to authorize requests to a non-Firebase backend. Requests to Firebase
+   * backends are authorized automatically if configured.
+   */
+  Future<AppCheckToken> GetAppCheckToken(bool force_refresh) = 0;
+
+  /**
+   * Registers an {@link AppCheckListener} to changes in the token state. This
+   * method should be used ONLY if you need to authorize requests to a
+   * non-Firebase backend. Requests to Firebase backends are authorized
+   * automatically if configured.
+   */
+  void AddAppCheckListener(AppCheckListener* listener) = 0;
+
+  /** Unregisters an {@link AppCheckListener} to changes in the token state. */
+  void RemoveAppCheckListener(AppCheckListener* listener) = 0;
+}
 
 #endif  // FIREBASE_APP_CHECK_SRC_INCLUDE_FIREBASE_APP_CHECK_H_
 
 }  // namespace app_check
 }  // namespace firebase
+
